+++ CHANGELOG.md
@@ -294,3 +294,6 @@
 * Define `substructure` expression
 * Define `o'rw'o` operator
 * Define `ds` operator
+* Define `Fallible` type alias
+* Define `Failure` pattern synonym
+* Define `Success` pattern synonym